pyproject.toml
[project]
name = "mcpterm"
version = "0.1.0"
description = "a cli for AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.7.1",
    "ollama>=0.4.8",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "sentence-transformers>=4.1.0",
    "sqlite-vec>=0.1.6",
]
